git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11361 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/mission.cpp
@@ -781,7 +781,7 @@ while (MsnGetS (buf, 80, cf)) {
 				PrintLog (0, "'%s'\n", buf);
 				MsnTrimComment (buf);
 				MsnAddStrTerm (buf);
-				if (strlen (buf) > (nEnhancedMission ? 255 : 12)) {
+				if ((int) strlen (buf) > (nEnhancedMission ? 255 : 12)) {
 					PrintLog (-1, "mission file: invalid level name\n");
 					return 0;
 					}
